feat: 🎸 Port MFA to ts
